+++ components/scg/src/command/get-node-from-memory.js
@@ -16,7 +16,8 @@ SCgCommandGetNodeFromMemory.prototype = {
         this.object.setText(this.oldIdtf);
         this.object.setScType(this.oldType);
         if (this.oldScAddr != null) {
-            this.object.setScAddr(this.oldScAddr, true);
+            this.object.setScAddr(this.oldScAddr);
+            this.object.setObjectState(SCgObjectState.MergedWithMemory);
         } else {
             this.object.sc_addr = null;
             this.object.state = SCgObjectState.Normal;
@@ -27,7 +28,8 @@ SCgCommandGetNodeFromMemory.prototype = {
 
     execute: function () {
         this.object.setText(this.newIdtf);
-        this.object.setScAddr(this.newScAddr, true);
+        this.object.setScAddr(this.newScAddr);
+        this.object.setObjectState(SCgObjectState.MergedWithMemory);
         this.object.setScType(this.newType);
         this.scene.updateRender();
     }